nothing of note

+++ src/main.cc
@@ -119,7 +119,7 @@ int main(int argc, char* argv[]){
             }
 
             char* invalid; //invalid/valid pointer
-            int cd_loc = strtol(input.c_str(), &invalid, 10); // converts string to int
+            int cd_loc = (int)strtol(input.c_str(), &invalid, 10); // converts string to int
             if(*invalid){ // if input is invalid
                 std::cout << "\tInvalid number.\n"; 
                 break;
